% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/school_palettes.R
\name{school_palette}
\alias{school_palette}
\title{A School Palette Generator}
\usage{
school_palette(
  name,
  n,
  type = c("all", "primary", "secondary"),
  scale = c("discrete", "continuous")
)
}
\arguments{
\item{name}{Name of desired palette.}

\item{n}{Number of colors desired. Many schools only have only a handful of
colors. If omitted, uses all colors.}

\item{type}{One of "all", "primary", or "secondary". "all" will return all
possible colors, "primary" will return just the main colors for a school,
and "secondary" will return all other available colors other than "primary".}

\item{scale}{Either "discrete" or "continuous". Use continuous if you want
to automatically interpolate between colors.}
}
\value{
A vector of colors.
}
\description{
These are the color palettes from NCAA schools.
}
\keyword{colors}
