% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/school_palettes.R
\docType{data}
\name{school_palettes}
\alias{school_palettes}
\title{Complete List of School Palettes}
\format{
An object of class \code{list} of length 1159.
}
\usage{
school_palettes
}
\description{
Use \code{\link{school_palette}} to construct palettes of desired length.
The list of schools is pulled from \href{https://www.ncaa.com/schools-index}{NCAA Schools}
}
\keyword{datasets}
