% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schoolcolors-package.R
\docType{package}
\name{schoolcolors-package}
\alias{schoolcolors}
\alias{schoolcolors-package}
\title{schoolcolors: Color Palettes for Univeristies}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dermcnor/schoolcolors}
  \item Report bugs at \url{https://github.com/dermcnor/schoolcolors/issues}
}

}
\author{
\strong{Maintainer}: Derek Norton \email{dnorton1@gmail.com}

Authors:
\itemize{
  \item Cris Benge \email{crisbenge@hotmail.com}
}

}
\keyword{internal}
